# Optimize order by single nullable or low-cardinality columns