@Algunenano 2025-03-18
`group_by_fixed_keys	5` has slowed down (`+1.465x`) but I don't see how that's possible, and running it locally I see the same perf.