@taiyang-li 2025-03-19
> `group_by_fixed_keys 5` has slowed down (`+1.465x`) but I don't see how that's possible, and running it locally I see the same perf.

Interesting, I'll look into it.