@taiyang-li 2025-03-19
@Algunenano I can't reproduce it in my local environment.

```
./clickhouse benchmark --concurrency 1  -i 20   --max_threads=1 --query "select a, b from group_by_fk group by a, b format Null"

Before
localhost:9001, queries: 20, QPS: 1.010, RPS: 10096806.800, MiB/s: 77.033, result RPS: 0.000, result MiB/s: 0.000.
0%              0.756 sec.
10%             0.960 sec.
20%             0.969 sec.
30%             0.978 sec.
40%             0.979 sec.
50%             0.983 sec.
60%             0.988 sec.
70%             0.993 sec.
80%             0.999 sec.
90%             1.008 sec.
95%             1.010 sec.
99%             1.343 sec.
99.9%           1.343 sec.
99.99%          1.343 sec.

After
localhost:9001, queries: 20, QPS: 0.995, RPS: 9948095.166, MiB/s: 75.898, result RPS: 0.000, result MiB/s: 0.000.
0%              0.861 sec.
10%             0.974 sec.
20%             0.975 sec.
30%             0.979 sec.
40%             0.986 sec.
50%             0.999 sec.
60%             1.000 sec.
70%             1.015 sec.
80%             1.036 sec.
90%             1.046 sec.
95%             1.049 sec.
99%             1.120 sec.
99.9%           1.120 sec.
99.99%          1.120 sec. 
```